pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voxsentinel"
version = "0.1.0"
description = "VoxSentinel — Real-time telephony transcription & compliance guard"
requires-python = ">=3.11"

# All runtime + dev dependencies unified here so a single venv serves
# every package and service during local development and CI.
dependencies = [
    # ── tg-common runtime ──
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "redis>=5.0",
    "celery>=5.4",
    "structlog>=24.2",
    "prometheus-client>=0.20",

    # ── ingestion runtime ──
    # "av>=12.0",        # requires FFmpeg; mocked in unit tests
    "numpy>=1.26",

    # ── vad runtime ──
    # "torch>=2.3",      # heavy; mocked in unit tests

    # ── asr runtime ──
    "circuitbreaker>=2.0",
    # "deepgram-sdk>=3.0",   # heavy; mocked in unit tests
    # "faster-whisper>=1.1",  # requires CTranslate2/GPU; mocked in unit tests

    # ── nlp runtime ──
    "pyahocorasick>=2.1",
    "rapidfuzz>=3.8",
    # "transformers>=4.40",       # heavy ML; mocked in unit tests
    # "presidio-analyzer>=2.2",   # heavy NLP; mocked in unit tests
    # "presidio-anonymizer>=2.2", # heavy NLP; mocked in unit tests
    # "spacy>=3.7",               # heavy; mocked in unit tests
    # "torch>=2.3",               # heavy; already mocked

    # ── shared service deps ──
    "fastapi>=0.111",
    "uvicorn>=0.30",
    "httpx>=0.27",
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-mock>=3.14",
    "ruff>=0.4",
    "mypy>=1.10",
    "testcontainers[postgres,redis,elasticsearch]>=4.4",
    "gtts>=2.5",
    "pydub>=0.25",
    "elasticsearch[async]>=8.13",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true
explicit_package_bases = true
exclude = ["tests/", "conftest\\.py$"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--import-mode=importlib"
# Test discovery covers every sub-project from the root.
testpaths = [
    "packages/tg-common/tests",
    "services/ingestion/tests",
    "services/vad/tests",
    "services/asr/tests",
    "services/nlp/tests",
    "services/diarization/tests",
    "services/alerts/tests",
    "services/storage/tests",
    "services/api/tests",
    "tests/integration",
]
markers = [
    "integration: marks tests as integration tests (require Docker)",
]
filterwarnings = [
    # testcontainers' RedisContainer still uses the deprecated @wait_container_is_ready
    # decorator internally — we already bypass it with our own TCP readiness check.
    "ignore::DeprecationWarning:testcontainers.redis",
]
